move views related to permissions (Actions & Permissions)
